move to setup.py
